requirements.txt
streamlit==1.32.2
pinecone-client==3.2.2
langchain==0.1.16
langchain-google-genai==1.0.3
google-generativeai==0.5.2
tiktoken==0.6.0
pdfplumber==0.10.3